fix: rename prometheus pkg to metrics and change endpoints
