--- pkg/prometheus/cleanup.go
@@ -1,9 +1,0 @@
-package prometheus
-
-import "github.com/prometheus/client_golang/prometheus"
-
-func Cleanup(mm *MetricsManager) func() {
-	return func() {
-		prometheus.DefaultRegisterer.Unregister(mm.prometheusMetrics)
-	}
-}
